Move classes to cli folder to be part of the package
