+++ cli/.npmignore
@@ -1,2 +1,3 @@
 pokemon-names.json
 tasks.json
+assets/cli.gif
