+++ cli/item-manager.mjs
